Hydra.Client

Documentation

data HydraEvent tx

Constructors

ClientConnected

ClientDisconnected

Update (ServerOutput tx)

newtype Client tx m

Handle to provide a means for sending inputs to the server.

Constructors

Client

Fields

sendInput :: ClientInput tx -> m ()

Send some input to the server.

type ClientCallback tx m = HydraEvent tx -> m ()

Callback for receiving server outputs.

type ClientComponent tx m a = ClientCallback tx m -> (Client tx m -> m a) -> m a

A type tying both receiving output and sending input into a Component.

withClient :: Tx tx => Host -> ClientComponent tx IO a

Connect to a hydra-node using a websocket and provide a Component for sending client inputs, as well as receiving server outputs.

data ClientError

Constructors

ClientJSONDecodeError String ByteString